requirements.txt
aiohttp~=3.7.4.post0
SQLAlchemy~=1.4.22
python-dotenv~=0.19.0
aiopg~=1.3.1